organized
I put everything i needed into folders.
I am going to double check everything.

After, I am going to submit this just so it is not late. Then, I am going to polish as much as I can and fix errors I was not able to fix before. If I do not polish in time, at least I have it handed in.