GitHub Issue - Possible Fix
My github the last second stopped working. When I went to do it online instead of through desktop, it was not allowing me to create folders without adding a random file in it (which I do not understand how. I might have done it wrong)
To save time and not go through the trouble of slowly creating folder after folder, the file after file. I just added all of them to my index branch since it is working.